feat(jc): unified v2 production flow + auth positions
Backend
- Migrations 009-025: planning_v2 chain, JC v2 sub-tables, accounting,
  annexure v2 (5 types), app-supplied 8-digit IDs, PM indent relocation
  to stage 1, consumed_qty + bom_line_id on indent + consumption,
  so_upload + production_head roles seeded.
- Services: job_card_v2, plan_v2, jc_annexures_v2, jc_accounting_v2,
  fulfillment_v2, uom; per-stage Material Consumption (RM + PM),
  BOM-catalog dropdowns surfaced on every JC.
- REQUIRED_SIGN_OFFS = ('production_head',); legacy roles still seeded
  for audit but no longer gate close.
- dispatch-to-next rejects qty_kg <= 0 at Pydantic (gt=0).
- acknowledge-material persists notes as v2 remark.
- Sign-off SignOffRequest accepts signed_by_name (typed name overrides
  JWT signer); record-output v2 accepts rm_consumed[] + pm_consumed[]
  per-line with bom_line_id; receive-material per-box savepoints +
  FOR UPDATE on indent row; soft-delete FOR UPDATE.
- Password complexity rules removed (evaluate() returns [] always);
  bcrypt hashing unchanged.

+++ app/db/migrate.sql
@@ -119,7 +119,7 @@ $$;
 -- 7. Create GST reconciliation table
 CREATE TABLE IF NOT EXISTS so_gst_reconciliation (
     recon_id              SERIAL PRIMARY KEY,
-    so_line_id            INT NOT NULL REFERENCES so_line(so_line_id),
+    so_line_id            BIGINT NOT NULL REFERENCES so_line(so_line_id),
     so_id                 INT NOT NULL REFERENCES so_header(so_id),
     expected_gst_rate     NUMERIC,
     actual_gst_rate       NUMERIC,
